requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
PyPDF2
textstat
reportlab
transformers
torch
sentencepiece
plotly
pymongo
requests
copyleaks
copyleaks
python-dotenv
requests
